Lute/removing clingoext (#19)
* removing clingoext file
