+++ helper_test/helper_clingo.py
@@ -1,4 +1,5 @@
-from eclingo.util import clingoext as _clingoext
+import eclingo.internal_states.internal_control as internal_control
+
 from eclingo.util.logger import silent_logger
 from . import helper
 
@@ -8,7 +9,7 @@ class ClingoTestHelper(helper.TestHelper):
 
     def setUp(self):
         super().setUp()
-        self.clingo_control = _clingoext.Control(logger=silent_logger)
+        self.clingo_control = internal_control.InternalStateControl(logger=silent_logger)
         self.program_added = False
 
     def add_program(self, program):
