+++ helper_test/helper_parsing.py
@@ -4,12 +4,12 @@ import clingo as _clingo
 
 import clingo.ast as _ast
 
-from eclingo.util import clingoext as _clingoext
 from eclingo.util.logger import silent_logger
 from eclingo.util.astutil import ast_repr as _ast_repr
 
 from eclingo.parsing.parser import parse_program as _parse_program
 from eclingo.internal_states import ShowStatement
+
 
 from . import helper
