+++ noxfile.py
@@ -2,8 +2,14 @@ import nox
 
 # If using conda, then be sure to be in the nox environment different from "base" before running this file.
 
+def typecheck(session):
+    session.install("-r", "requirements.txt", "mypy")
+    session.run("mypy", "src/eclingo")
+
 @nox.session(python=None)
 def tests(session):
     args = session.posargs or ["--cov"]
-    session.run("poetry", "install", external=True)
+    # session.run("poetry", "install", external=True)
+    session.install("-r", "requirements.txt")
+    session.install("-e", ".")
     session.run("pytest", *args)
