+++ src/eclingo/internal_states/internal_control.py
@@ -1,21 +1,27 @@
 from abc import abstractmethod
 from dataclasses import dataclass
-from typing import Any, Callable, Iterable, Iterator, Sequence, Tuple, Union
+import textwrap
+from typing import Any, Callable, Dict, Iterable, Iterator, List, Sequence, Tuple, Union
 
-import clingo as _clingo
+import clingo
 from clingo import MessageCode, Symbol, SymbolicAtom
-from clingo import ast as _ast
+from clingo import ast
+from clingo.ast import parse_string
 
+
+from eclingo.util import astutil
 from eclingo.prefixes import atom_user_name
-from eclingo.util import clingoext as _clingoext
 
 from .mappings import EpistemicSymbolToTestSymbolMapping, SymbolToEpistemicLiteralMapping, SymbolToEpistemicLiteralMappingUsingProgramLiterals, SymbolToEpistemicLiteralMappingUsingShowStatements
 
+import clingox
+from clingox import program as clingox_program
+from clingox.backend import SymbolicBackend
 
 class ASTParsedObject():
     pass
 
-ASTObject = Union[ASTParsedObject, _ast.AST]  # pylint: disable=no-member
+ASTObject = Union[ASTParsedObject, ast.AST]  # pylint: disable=no-member
 
 
 @dataclass(frozen=True)
@@ -28,44 +34,106 @@ class ShowSignature(set):
     pass
 
 
-class ProgramBuilder(_clingoext.ProgramBuilder):
-
+class ProgramBuilder():
+            
     def __init__(self, control, program, show_signature):
-        super().__init__(control, program)
+        self.control = control
+        self.program = program
         self.show_signature = show_signature
-
+        self.bulider = clingo.ast.ProgramBuilder(self.control)
+        
     def add(self, statement: ASTObject):
         if isinstance(statement, ShowStatement):
             self.show_signature.add(statement)
-        elif isinstance(statement, _ast.AST): # pylint: disable=no-member
-            return super().add(statement)
+        elif isinstance(statement, ast.AST): # pylint: disable=no-member
+            return self.add_st(statement)
         else:
             raise RuntimeError("Non recognised object: " + str(statement))
 
+    def __enter__(self):
+        self.bulider.__enter__()
+        return self
 
-class InternalStateControl(_clingoext.Control):
+    def __exit__(self, type_, value, traceback):
 
-    def __init__(self, arguments: Sequence[str] = (), logger: Callable[[MessageCode, str], None] = None, message_limit: int = 20, *, control: _clingo.Control = None):
-        super().__init__(arguments, logger, message_limit, control=control)
-        self.show_signature = ShowSignature()
+        return self.bulider.__exit__(type_, value, traceback)
+
+    def add_st(self, statement: ast.AST): # pylint: disable=no-member
+        self.program.append(statement)
+        try:
+            return self.bulider.add(statement)
+        except RuntimeError as error:
+            if len(error.args) != 1:
+                raise error
+            if error.args[0] == 'literal expected':
+                error.args = ('literal expected, got\n' + textwrap.indent(astutil.ast_repr(statement), 13*' '), )
+            raise error
+        except AttributeError as error:
+            if error.args[0] == "'list' object has no attribute 'location'":
+                error.args = (error.args[0] + '\n' + textwrap.indent(astutil.ast_repr(statement), 13*' '), )
+            raise error     
+        
+
+
+class InternalStateControl(object):
+
+    def __init__(self, arguments: Sequence[str] = (), logger: Callable[[MessageCode, str], None] = None, message_limit: int = 20, *, control: clingo.Control = None):
+        
+        if control is None:
+            control = clingo.Control(arguments, logger, message_limit)
+        self.control = control
+        
+        self.parsed_program: List[ast.AST] = [] # pylint: disable=no-member
+              
+        self.ground_program = clingox_program.Program()
+        self.control.register_observer(clingox_program.ProgramObserver(self.ground_program))
+        
+        # Legacy
+        self.show_signature = ShowSignature() # What is this supposed to be doing?
+        
         self.epistemic_to_test_mapping = EpistemicSymbolToTestSymbolMapping()
         self.show_mapping = SymbolToEpistemicLiteralMapping()
+        
+    def add_program(self, program: str) -> None:
+        with self.builder() as builder:
+            parse_string(program, builder.add)
 
+    # Remove self.show_signature from arguments (?)
     def builder(self) -> ProgramBuilder:
         return ProgramBuilder(self.control, self.parsed_program, self.show_signature)
 
+    def add_to(self, control: Union['InternalStateControl', clingo.Control]):
+        program = self.ground_program
+        with control.backend() as backend:
+            mapping = clingox_program.Remapping(backend, program.output_atoms, program.facts)
+            program.add_to_backend(backend, mapping)
+        return mapping
+
+    def facts(self) -> Iterable[Symbol]:
+        for symbolic_atom in self.control.symbolic_atoms:
+            if symbolic_atom.is_fact:
+                yield symbolic_atom.symbol
+                
+    def show_symbols(self) -> Iterator[Symbol]:
+        for symbolic_atom in self.show_symbolic_atoms():
+            yield symbolic_atom.symbol
+
+    def atom_to_symbol_mapping(self) -> Dict[int, Symbol]:
+        mapping = dict()
+        for symbolic_atom in self.control.symbolic_atoms:
+            if not symbolic_atom.is_fact:
+                mapping.update({symbolic_atom.literal : symbolic_atom.symbol})
+        return mapping
+    
     def show_symbolic_atoms(self) -> Iterator[SymbolicAtom]:
         for show_statement in self.show_signature:
             symbolic_atoms = self.control.symbolic_atoms
             show_statment_user_name = atom_user_name(show_statement.name)
             yield from symbolic_atoms.by_signature(show_statment_user_name, show_statement.arity, show_statement.poistive)
 
-    def show_symbols(self) -> Iterator[Symbol]:
-        for symbolic_atom in self.show_symbolic_atoms():
-            yield symbolic_atom.symbol
-
+    
     def ground(self, parts: Sequence[Tuple[str, Sequence[Symbol]]], context: Any = None) -> None:
-        super().ground(parts, context)
+        self.control.ground(parts, context)
         self.epistemic_to_test_mapping = EpistemicSymbolToTestSymbolMapping(self)
         self.show_mapping = self._generate_show_mapping()
 
@@ -74,8 +142,15 @@ class InternalStateControl(_clingoext.Control):
             return SymbolToEpistemicLiteralMappingUsingShowStatements(self.show_symbols())
         else:
             return SymbolToEpistemicLiteralMappingUsingProgramLiterals(self.epistemic_to_test_mapping.epistemic_literals())
+    
 
+    def symbolic_backend(self) -> SymbolicBackend:
+        return clingox.backend.SymbolicBackend(self.control.backend())
 
+    def __getattr__(self, attr):
+        if attr in self.__dict__:
+            return getattr(self, attr)
+        return getattr(self.control, attr)
 
 
 class Application(object):
@@ -85,9 +160,11 @@ class Application(object):
         raise NotImplementedError
 
 
-class ApplicationWrapper(_clingoext.ApplicationWrapper):
+class ApplicationWrapper(clingo.Application):
+    def __init__(self, application):
+        self.application = application
 
-    def main(self, control: _clingo.Control, files: Sequence[str]) -> None:
+    def main(self, control: clingo.Control, files: Sequence[str]) -> None:
         internal_control = InternalStateControl(control=control)
         return self.application.main(internal_control, files)
 
@@ -99,4 +176,4 @@ class ApplicationWrapper(_clingoext.ApplicationWrapper):
 
 def clingo_main(application: Application, files: Sequence[str] = ()) -> int:
     application_wrapper = ApplicationWrapper(application)
-    return _clingo.clingo_main(application_wrapper, files)
+    return clingo.clingo_main(application_wrapper, files)
