+++ src/eclingo/internal_states/mappings.py
@@ -1,16 +1,21 @@
-from typing import Dict, Iterable, MutableMapping, Optional  # , Dict
+from typing import Dict, Iterable, Union, TYPE_CHECKING
 
 from clingo import Function, Symbol
 from clingo.ast import Sign #pylint: disable=import-error
 
 from eclingo import prefixes
-from eclingo.literals import EpistemicLiteral, Literal
-from eclingo.util import clingoext
 
+from eclingo.literals import EpistemicLiteral, Literal
+import clingo
+
+if TYPE_CHECKING:
+    # This takes care of a cyclic import and it is only executed a typechecking time
+    # It would be nice to find a better way to handle this
+    from eclingo.internal_states.internal_control import InternalStateControl
 
 class EpistemicSymbolToTestSymbolMapping(dict[Symbol, Symbol]):
 
-    def __init__(self, control: Optional[clingoext.Control] = None) -> None:
+    def __init__(self, control: Union[clingo.control.Control, 'InternalStateControl', None] = None) -> None:
         super().__init__()
         if control is not None:
             for atom in control.symbolic_atoms:
