+++ tests/test_clingoext.py
@@ -3,9 +3,9 @@ import clingo
 from clingo import Function, Number
 from clingo import ast as _ast
 
-import eclingo.util.clingoext as clingoext
 from eclingo.util.groundprogram import *
 from eclingo.util.logger import silent_logger
+import eclingo.internal_states.internal_control as internal_control
 
 from clingo.ast import  parse_string, Location, Position
 
@@ -20,7 +20,7 @@ f  = clingo.Function('f', [], True)
 class Test(unittest.TestCase):
 
     def setUp(self):
-        self.control = clingoext.Control(logger=silent_logger)
+        self.control = internal_control.InternalStateControl(logger=silent_logger)
 
     def assert_models(self, models, obtained_models):
         models.sort()
